% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_basemap}
\alias{add_basemap}
\title{Add a Polygon Basemap}
\usage{
add_basemap(data = mvldata::nz_tla_06.df, fill = mvl_half_grey,
  colour = mvl_grey, size = 0.2, hole_fill = "white")
}
\arguments{
\item{data}{Usually will be mvldata::nz_tla_06.df, nz_tla_13.df, or nz_regions.df; but can be any data.frame of polygons with [long, lat, group] attributes}

\item{fill}{colour to fill the polygons}

\item{colour}{colour of the polygon borders}

\item{size}{width of the border line}
}
\description{
Used to quickly add a basemap of NZ geographical areas to plots
}
