% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_placenames}
\alias{add_placenames}
\title{Add Some Placenames to a Map}
\usage{
add_placenames(data = mvldata::sparse_places.df, colour = mvl_text,
  fillcolour = "white", size = 3, tla = NULL)
}
\arguments{
\item{data}{A data.frame of the names to add to the plot, and their locations in NZTM coordinates}

\item{colour}{The text colour for the place names}

\item{fillcolour}{Fill colour for the text. Only used if shadowtext package is installed.}

\item{size}{The text size for the place names}

\item{tla}{Optional argument to only show the suburbs of a given set of TLA ids (2006 TLA definition)}
}
\description{
Use to quickly add a layer of towns/cities and some selected suburbs to a map for reference
}
