% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_water}
\alias{add_water}
\title{Add Water Polygons to a Map}
\usage{
add_water(rivers = TRUE, inland = TRUE, extra = TRUE,
  fill = mvl_half_teal, colour = mvl_teal, size = 0.2,
  hole_fill = mvl_half_grey)
}
\arguments{
\item{rivers}{Should river polygons be added to the plot?}

\item{inland}{Should inland water polygons be added to the plot?}

\item{extra}{Should additional lakes/rivers not covered by inland and rivers be added?}

\item{fill}{colour to fill the polygons}

\item{colour}{colour of the polygon borders}

\item{size}{width of the border line}

\item{hole_fill}{what colour should holes be filled with?}
}
\description{
Used to quickly add water polygons to plots
}
