% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{bcomma}
\alias{bcomma}
\title{Label in billions}
\usage{
bcomma(x, dp = 1, form = "f", ...)
}
\arguments{
\item{x}{A numeric amount}
}
\value{
A character representation of x in billions
}
\description{
A modification of the \code{\link[scales]{comma}} function from \code{scales}, which
is more appropriate for plots of large values
}
\examples{
bcomma(1000000000)
bcomma(c(2000000000, 15000000000))
}
