% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{bdollar}
\alias{bdollar}
\title{Label in billions of dollars}
\usage{
bdollar(x, dp = 2, form = "f", ...)
}
\arguments{
\item{x}{A numeric dollar amount}
}
\value{
A character representation of x in billions of dollars
}
\description{
A modification of the \code{dollar()} function from \code{scales}, which
is more appropriate for plots of large dollar values
}
\examples{
bdollar(1000000000)
bdollar(c(2000000000, 15000000000))
}

