% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{bnz_spend_origin}
\alias{bnz_spend_origin}
\title{Get the origin of BNZ spending for a given merchant.}
\usage{
bnz_spend_origin(db_con, group_id,
  seqmonth_start = as.numeric(format(min(seq.Date(from = Sys.Date(), by =
  "-1 months", length.out = 13)), "\%Y\%m")),
  seqmonth_end = as.numeric(format(seq.Date(from = Sys.Date(), by =
  "-1 months", length.out = 13)[2], "\%Y\%m")))
}
\arguments{
\item{db_con}{A connection to the BNZ database}

\item{group_id}{The group ID(s) of the merchant(s) of interest}

\item{seqmonth_start}{The first month from which to get spending. Defaults to 13 months prior to Sys.Date()}

\item{seqmonth_end}{The last month from which to get spending. Defaults to one month prior to Sys.Date()}
}
\description{
Given a group ID and a time frame, \code{bnz_spend_origin()} returns a data frame containing the spending by customer CAU.
}
\details{
There might be some issues with the default date ranges if the current date is at the start or end of the month, but it
should be good enough and you can set it manually if you're concerned.
}
\examples{
DB <- odbcConnect("MVIEW", uid="user", pwd="password")
bnz_spend_origin(DB, group_id = 15435)
}
