% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{bounding_box}
\alias{bounding_box}
\title{Add a bounding box to a map plot.}
\usage{
bounding_box(long_limits, lat_limits, draw_outline = TRUE)
}
\arguments{
\item{long_limits}{The longitude range that the map should take}

\item{lat_limits}{The latitude range that the map should take}

\item{draw_outline}{Should the border of the box be drawn?}
}
\value{
Adds a box to the outline of the plot sets the plot limits
}
\description{
Add a bounding box to a ggplot map and set the limits of the plot to the box.
}

