% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{cagr}
\alias{cagr}
\title{Compound Annual Growth Rate}
\usage{
cagr(iv, fv, length)
}
\arguments{
\item{iv}{Initial value}

\item{fv}{Final value}

\item{length}{Length of time between \code{iv} and \code{fv}}
}
\value{
cagr The compound annual growth rate from \code{iv} to \code{fv}
}
\description{
Compound Annual Growth Rate
}
\examples{
sales <- data.frame(year_1 = c(10, 5, 1), year_2 = c(15, 4, 5), year_3 = c(20, 8, 2))
cagr(iv = sales$year_1, fv = sales$year_3, length = 2)
}

