% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calculate.tm.p.values}
\alias{calculate.tm.p.values}
\title{Calculate p-values for a tranformation matrix}
\usage{
calculate.tm.p.values(tm.obs, tm.null, method = "z-score")
}
\arguments{
\item{tm.obs}{The observed transition matrix}

\item{tm.null}{A list of null transition matrices}

\item{method}{one of 'z-score' or 'non-parametric'}
}
\description{
This function calculates the significance of an observed transition matrix given a set of null transition matrices
}
\examples{
example1
}
\keyword{keywords}

