% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_fwhen.R
\name{case_fwhen}
\alias{case_fwhen}
\title{Vectorised if returning an ordered result}
\usage{
case_fwhen(...)
}
\arguments{
\item{...}{A sequence of two-sided formulas. The left hand side (LHS)
determines which values match this case. The right hand side (RHS)
provides the replacement value.

The LHS must evaluate to a logical vector. The RHS does need to be
logical, but all RHSs must evaluate to the same type of vector.

Both LHS and RHS may have the same length of either 1 or \code{n}. The
value of \code{n} must be consistent across all cases. The case of
\code{n == 0} is treated as a variant of \code{n != 1}.

These dots support \link[rlang:tidy-dots]{tidy dots} features.}
}
\description{
This function allows you to vectorise multiple if and else if statements. It is an R equivalent of the SQL CASE WHEN statement.
An update from the dplyr implementation as it returns the result as an ordered factor in the order given by the formulae
}
