% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_picker.R
\name{catchment_picker}
\alias{catchment_picker}
\title{Catchment selection tool.}
\usage{
catchment_picker(spending, location, map_shapefile = mvldata::nz_cau_13.spdf)
}
\arguments{
\item{spending}{A data.frame of the spend origin with the attributes CAU and SPEND}

\item{location}{A data.frame with the merchant coordinates as X_COORD, Y_COORD in NZTM projection}

\item{map_shapefile}{A shapefile from which to draw the map (default is 2013 CAU)}
}
\description{
A Shiny function to make picking a catchment for a merchant or retail area easier.
}
