% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{comma_change}
\alias{comma_change}
\title{Label with '+' for positive growth}
\usage{
comma_change(x, dp = 0, form = "f", ...)
}
\arguments{
\item{x}{A numeric amount}
}
\value{
A character representation of x in dollars
}
\description{
A modification of the \code{\link[scales]{comma}} function from \code{scales}, which
is more appropriate for movements
}
\examples{
comma_change(1000000)
comma_change(c(-2000000, 115000000))
}
