% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_post_campaign.R
\name{create_mcd_run}
\alias{create_mcd_run}
\title{Create a run script.}
\usage{
create_mcd_run(campaign_name, root_dir)
}
\arguments{
\item{campaign_name}{The name of the R project.}

\item{root_dir}{The root directory of the R project.}
}
\value{
Creates a script within the directory specified.
}
\description{
Builds the skeleton of a run script in the specified root directory.
Primarily for internal use in create_project().
}
