% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\alias{create_mcd_project}
\title{Create a new R Project.}
\usage{
create_project(project_name, base_dir = "M:/clients", within = NULL,
  sub_dirs = NULL)

create_mcd_project(project_name,
  base_dir = "M:/clients/mcdonalds/2018/Business Insights/Contract Work/Data Requests/",
  within = NULL, sub_dirs = NULL)
}
\arguments{
\item{project_name}{The name that the project will take}

\item{base_dir}{The base directory for the project}

\item{within}{Optional path to a subdirectory within base_dir where the project will be located}

\item{sub_dirs}{Optional names for any additional sub-direcories that the project should contain}
}
\value{
Logical, indicating that the project has been successfully created
}
\description{
Create some common sub-directories used in an R project, and import
some scripts containing headers for common analysis tasks.
}
