% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_rmd}
\alias{create_rmd}
\title{Create a new R Markdown script
Creates a new R Markdown script with the headers and workflow mapped out in the same
manner as a standard scrpt from create_project()}
\usage{
create_rmd(name, location = "/")
}
\arguments{
\item{name}{The name for the new RMd script}

\item{location}{Directory where the script should be created - default is the root directory}
}
\description{
Create a new R Markdown script
Creates a new R Markdown script with the headers and workflow mapped out in the same
manner as a standard scrpt from create_project()
}
