% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_script}
\alias{create_script}
\title{Create a new R script
Creates a new R script with the headers and workflow mapped out in the same
manner as a standard results scrpt from create_project()}
\usage{
create_script(name, project_name = getwd(), location = "R")
}
\arguments{
\item{name}{The name for the new R script}

\item{project_name}{Name of the project (should always be the current working directory)}

\item{location}{Directory where the script should be created - default is the /R directory}
}
\description{
Create a new R script
Creates a new R script with the headers and workflow mapped out in the same
manner as a standard results scrpt from create_project()
}
