% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{data_import}
\alias{data_import}
\title{Import the processed data for a project}
\usage{
data_import(base_dir = getwd(), data_name = paste(gsub(".*/(?!$)|/$", "",
  base_dir, perl = TRUE), "data.RData"))
}
\arguments{
\item{base_dir}{The base directory of the project}

\item{data_name}{The name that the data object has. Defaults to "project name data.RData"}
}
\value{
imports the processed data into the global environment
}
\description{
Import the processed data for a project
}
