% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{dollar}
\alias{dollar}
\title{Label in dollars}
\usage{
dollar(x, dp = 0, form = "f", ...)
}
\arguments{
\item{x}{A numeric dollar amount}
}
\value{
A character representation of x in dollars
}
\description{
A modification of the \code{\link[scales]{dollar}} function from \code{scales}, which
is more appropriate for negative values
}
\examples{
dollar(1000000)
dollar(c(-2000000, 15000000))
}
