% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporteRs functions.R
\name{flextable_black}
\alias{flextable_black}
\title{Create a black-themed FlexTable}
\usage{
flextable_black(df, size = 12, padding = size/2)
}
\arguments{
\item{df}{The data which should populate the flextable}

\item{size}{Which size should the resulting FlexTable's text be?}

\item{padding}{How much padding should there be around the text?}
}
\description{
A function to create a FlexTable object for use in MS Office type outputs via the ReporteRs package
}
\examples{
flextable_black(head(cars))
}
