% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporteRs functions.R
\name{flextable_negatives}
\alias{flextable_negatives}
\title{Highlight Negative Cell Values in Red}
\usage{
flextable_negatives(ftbl, col_names)
}
\arguments{
\item{ftbl}{A FlexTable object to format}

\item{col_names}{The names of the columns in which to highlight the negatives}
}
\value{
ftbl The original FlexTable with the formatting applied
}
\description{
A function to highlight any negative values in a FlexTable for given column names.
This should work regardless of the formatting - character values such as those returned by \code{marketview::dollar()} or
\code{marketview::percent()} will be dealt with if I've done this right.
}

