% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporteRs functions.R
\name{flextable_positives}
\alias{flextable_positives}
\alias{ft_positives}
\title{Highlight Increased Cell Values in Green}
\usage{
flextable_positives(ftbl, col_names = names(ftbl$vals))

ft_positives(ftbl, col_names = names(ftbl$vals))
}
\arguments{
\item{ftbl}{A FlexTable object to format}

\item{col_names}{The names of the columns in which to highlight the positives}
}
\value{
ftbl The original FlexTable with the formatting applied
}
\description{
A function to highlight any positive values in a FlexTable for given column names.
This requires values to be formatted using \code{marketview::dollar_change()} or
\code{marketview::percent_change()} (or similar).
}
