% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporteRs functions.R
\name{flextable_row_highlight}
\alias{flextable_row_highlight}
\alias{ft_row_highlight}
\title{Highlight given rows of a flextable}
\usage{
flextable_row_highlight(ft, rows, cols = 1:ft$numcol)

ft_row_highlight(ft, rows, cols = 1:ft$numcol)
}
\arguments{
\item{ft}{The flextable object to highlight}

\item{rows}{The index of the rows that should be highlighted}
}
\value{
ft A modified FlexTable with the specified rows highlighted in gold
}
\description{
A function to highlight particular rows of a FlexTable with gold borders
}
\examples{
df_ex <- head(cars)
ft_ex <- flextable_teal(df_ex)
ft_ex <- flextable_row_highlight(ft_ex, which(df_ex$dist == max(df_ex$dist)))
}
