% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporteRs functions.R
\name{flextable_widths}
\alias{flextable_widths}
\alias{ft_widths}
\title{Set the Column Widths of a FlexTable Object}
\usage{
flextable_widths(ft, col_widths = rep(1, ft$numcol), table_length = 21.42)

ft_widths(ft, col_widths = rep(1, ft$numcol), table_length = 21.42)
}
\arguments{
\item{ft}{a Flextable created by FlexTable() or marketview::flextable() functions}

\item{col_widths}{A vector of the relative widths that you would like the columns to take}

\item{table_length}{The total width of the FLextable in centimeters}
}
\description{
Set the Column Widths of a FlexTable Object
}
