% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{get_merchant_location}
\alias{get_merchant_location}
\title{Fetch the location of a merchant from Fozzie.}
\usage{
get_merchant_location(db_con, group_id)
}
\arguments{
\item{db_con}{A connection to the BNZ database}

\item{group_id}{The group ID of the merchant of interest}
}
\description{
Given a group ID, \code{get_merchant_location()} returns a data frame containing the coordinates of the merchant, if it is available on Fozzie.
}
\examples{
DB <- odbcConnect("MVIEW", uid="user", pwd="password")
get_merchant_location(DB, group_id = 15435)
}

