% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{get_meshblock_data}
\alias{get_meshblock_data}
\title{Import a meshblock shapefile}
\usage{
get_meshblock_data(census = 2013)
}
\arguments{
\item{census}{Which census meshblocks should be retrieved?}
}
\value{
SpatialPolygonsDataFrame describing the meshblock boundaries
}
\description{
A function to import a shapefile of meshblock boundaries for a given census definition.
}
\examples{
mb_map.df <- get_meshblock_data(2013)
}
