% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{ieo_radius_locator}
\alias{ieo_radius_locator}
\title{Find IEO Merchants Within a Given Radius}
\usage{
ieo_radius_locator(DB, x, y, rad, impute = T)
}
\arguments{
\item{DB}{An ODBC connection to bespoke}

\item{x}{X coordinate of the centroid}

\item{y}{Y coordinate of the centroid}

\item{rad}{The radius in which to find merchants}

\item{impute}{Should merchants with imputed location be included in the results?}
}
\value{
A \code{data.frame} of the Fozzie IDs and locations of all merchants within the given radius
}
\description{
A function to find all the IEO merchants within a radius (in m) of a given location (in NZTM).
Optionally include merchants with imputed location from MB centroid.
}
