% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{kcomma}
\alias{kcomma}
\title{Label in thousands}
\usage{
kcomma(x, dp = 0, form = "f", ...)
}
\arguments{
\item{x}{A numeric amount}
}
\value{
A character representation of x in thousands of dollars
}
\description{
A modification of the \code{\link[scales]{comma}} function from \code{scales}, which
is more appropriate for plots of large values
}
\examples{
kcomma(10000)
kcomma(c(2000, 1500000))
}
