% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{kdollar}
\alias{kdollar}
\title{Label in thousands of dollars}
\usage{
kdollar(x, dp = 0, form = "f", ...)
}
\arguments{
\item{x}{A numeric dollar amount}
}
\value{
A character representation of x in thousands of dollars
}
\description{
A modification of the \code{\link[scales]{dollar}} function from \code{scales}, which
is more appropriate for plots of large dollar values
}
\examples{
kdollar(10000)
kdollar(c(2000, 1500000))
}

