% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_nztm.R
\name{latlon_to_nztm}
\alias{latlon_to_nztm}
\title{Convert lat-long Coordinates to NZTM Projection}
\usage{
latlon_to_nztm(dat)
}
\arguments{
\item{dat}{}
}
\description{
Convert lat-long Coordinates to NZTM Projection
}
