% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{make_cau_df}
\alias{make_cau_df}
\title{Create a data.frame of area unit boundaries}
\usage{
make_cau_df(cau_map, caus, tol = 25, agg = FALSE, agg_name = "aggregated")
}
\arguments{
\item{mb_map.spdf}{shapefile of the area unit boundaries}

\item{mbs}{subset of the area units to get}
}
\description{
Given a CAU shapefile and a list of area units, make_cau_df selects the area units and fortifies them into a data.frame for plotting
}
