% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_footnote}
\alias{make_footnote}
\title{Add a Marketview footnote to a plot}
\usage{
make_footnote(footnote_text = paste(format(Sys.time(), "\%d \%b \%Y")),
  size = 0.7, color = "grey",
  mvl_png = "M:/mvl_office/logo/2012/Marketview_Logo_Horizontal.png")
}
\arguments{
\item{footnote_text}{The text for the footnote to display}

\item{size}{The size of the footnote text}

\item{color}{The colour of the footnote text}

\item{mvl_png}{The location of the Marketview logo}
}
\value{
Adds a footnote to the current plot
}
\description{
Add a Marketview footnote to a plot
}

