% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{make_mb_df}
\alias{make_mb_df}
\title{Create a data.frame of meshblock boundaries}
\usage{
make_mb_df(mb_map, mbs)
}
\arguments{
\item{mbs}{subset of the meshblocks to get}

\item{mb_map.spdf}{shapefile of the meshblock boundaries}
}
\description{
Given a meshblock shapefile and a list of meshblocks, make_mb_df selects the meshblocks and fortifies them into a data.frame for plotting
}
