% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{map_get}
\alias{map_get}
\title{Load a map for plotting}
\usage{
map_get(city = "Auckland", census_version = 2013,
  map_dir = "M:/R/map_data")
}
\arguments{
\item{city}{The city which the map will cover}

\item{census_version}{From which census definition should the polygons come from?}

\item{map_dir}{Where are the RData files located?}
}
\value{
\code{map.df} A data frame describing the polygons with the following atributes:
     (long, lat, group, CAU, CAU_NAME)
}
\description{
Load a data frame containing the polygons for a given city
}

