% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{mb_radius_locator}
\alias{mb_radius_locator}
\title{Find Meshblocks Within a Given Radius}
\usage{
mb_radius_locator(DB, x, y, rad)
}
\arguments{
\item{DB}{An ODBC connection to bespoke}

\item{x}{X coordinate of the centroid}

\item{y}{Y coordinate of the centroid}

\item{rad}{The radius in which to find meshblocks}
}
\description{
A function to find all the meshblocks within a radius (in m) of a given location (in NZTM).
}
