% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_post_campaign.R
\name{mcd_post_campaign}
\alias{mcd_post_campaign}
\title{Create a new R Project.}
\usage{
mcd_post_campaign(campaign_name,
  base_dir = "M:/clients/mcdonalds/2018/Business Insights/Contract Work/POST CAMPAIGN ANALYSES",
  within = NULL, sub_dirs = NULL)
}
\arguments{
\item{campaign_name}{The name that the project will take}

\item{base_dir}{The base directory for the project}

\item{within}{Optional path to a subdirectory within base_dir where the project will be located}

\item{sub_dirs}{Optional names for any additional sub-direcories that the project should contain}
}
\value{
Logical, indicating that the project has been successfully created
}
\description{
Create some common sub-directories used in an R project, and import
some scripts containing headers for common analysis tasks.
}
