% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_process.R
\name{mcd_process}
\alias{mcd_process}
\title{Process Raw McDonalds Data.}
\usage{
mcd_process(dat)
}
\arguments{
\item{dat}{A data.frame from the MCD_BP_FINAL_CARD_DATASET table}
}
\value{
A processed data frame for analysis
}
\description{
A wrapper function to automatically format data taken from mcd_bp_final_data_card_dataset.
The attribute IEO_SEGMENT is defined as:
case when a.bunch_id in (19150, 19151) then b.bunch_name else b.bunch_namex end as ieo_segment
}
