% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcda_process.R
\name{mcda_process}
\alias{mcda_process}
\title{Process Raw Australian IEO Data}
\usage{
mcda_process(dat)
}
\arguments{
\item{dat}{Raw data from the DB}
}
\value{
dat: original data frame with attributes cleaned/ordered and new groups added
}
\description{
Process Raw Australian IEO Data
}
