% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{mcomma}
\alias{mcomma}
\title{Label in millions}
\usage{
mcomma(x, dp = 1, form = "f", ...)
}
\arguments{
\item{x}{A numeric dollar amount}
}
\value{
A character representation of x in millions
}
\description{
A modification of the \code{\link[scales]{comma}} function from \code{scales}, which
is more appropriate for plots of large values
}
\examples{
mcomma(1000000)
mcomma(c(2000000, 15000000))
}
