% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{mdollar}
\alias{mdollar}
\title{Label in millions of dollars}
\usage{
mdollar(x, dp = 2, form = "f", ...)
}
\arguments{
\item{x}{A numeric dollar amount}
}
\value{
A character representation of x in millions of dollars
}
\description{
A modification of the \code{dollar()} function from \code{scales}, which
is more appropriate for plots of large dollar values
}
\examples{
mdollar(1000000)
mdollar(c(2000000, 15000000))
}

