% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\docType{data}
\name{mvl_colourpal}
\alias{mvl_blue}
\alias{mvl_citrus}
\alias{mvl_colourpal}
\alias{mvl_fire}
\alias{mvl_fuchsia}
\alias{mvl_green}
\alias{mvl_grey}
\alias{mvl_half_grey}
\alias{mvl_half_hay}
\alias{mvl_half_leaf}
\alias{mvl_half_purple}
\alias{mvl_half_teal}
\alias{mvl_hay}
\alias{mvl_leaf}
\alias{mvl_navy}
\alias{mvl_orange}
\alias{mvl_pewter}
\alias{mvl_plum}
\alias{mvl_purple}
\alias{mvl_red}
\alias{mvl_royal}
\alias{mvl_royal2}
\alias{mvl_sky}
\alias{mvl_stone}
\alias{mvl_teal}
\alias{mvl_text}
\title{Marketview palette.}
\format{An object of class \code{character} of length 1.}
\usage{
mvl_blue

mvl_pewter

mvl_grey

mvl_red

mvl_orange

mvl_green

mvl_teal

mvl_half_teal

mvl_navy

mvl_royal

mvl_purple

mvl_half_purple

mvl_half_grey

mvl_hay

mvl_half_hay

mvl_leaf

mvl_half_leaf

mvl_stone

mvl_citrus

mvl_sky

mvl_plum

mvl_text

mvl_fire

mvl_royal2

mvl_fuchsia
}
\description{
Colours described by the \href{file:///M:/mvl_office/templates/Logos/MarketView_ColourGuide_v2.pdf}{Marketview colour palette}.
}
\keyword{datasets}

