% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_init.R
\name{mvl_init}
\alias{mvl_init}
\title{Initialise the Workspace for Analysis}
\usage{
mvl_init()
}
\description{
Essentially a function version of the standard MVL project initialise script. Will likely replace this script in the future.
}
