% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_process.R
\name{mvl_process}
\alias{mvl_process}
\title{Process Raw Marketview Data.}
\usage{
mvl_process(dat)
}
\arguments{
\item{dat}{A data.frame from bespoke}
}
\value{
A processed data frame for analysis
}
\description{
A function to format the standard analytics attributes which are stored as numeric variables to save table space
}

