% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{output_archive}
\alias{output_archive}
\title{Archive previous results from a project}
\usage{
output_archive(base_dir)
}
\arguments{
\item{base_dir}{The base directory of the project}
}
\description{
Archive previous results from a project
}

