% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{parse_seqmonth}
\alias{parse_seqmonth}
\title{Parse SEQMONTH attributes as dates}
\usage{
parse_seqmonth(x)
}
\arguments{
\item{x}{A vector of SEQMONTH}
}
\description{
Parse SEQMONTH attributes as dates
}
