% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{pca.plot}
\alias{pca.plot}
\title{Principal Components plot of transformation matrix}
\usage{
pca.plot(tm, title = "PCA Plot of Transition", clusters = 1, alpha = 1)
}
\arguments{
\item{tm}{a transition matrix for a bipartite network}

\item{title}{The title of the plot}

\item{clusters}{A vector indicating the colors to be plotted for each node}

\item{alpha}{A vector indicating the level of transparency to be plotted for each node}
}
\description{
This function plots the first two principal components for a transaction matrix
}
\examples{
data(yeast.panda)
t.matrix <- transformation.matrix(yeast.panda$cell.cycle, yeast.panda$stress.response)
p.values <- runif(ncol(t.matrix))  # Generate a uniform random to simulate p.values
clusters <- kmeans(t.matrix,3)$cluster # Color the nodes according to cluster membership
pca.plot(t.matrix,title="PCA Plot of Transition - Cell Cycle vs Stress Response", clusters=clusters,alpha=p.values)
}
\keyword{keywords}

