% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{percent_change}
\alias{percent_change}
\title{Convert numeric values to percentages, with '+' indicating positive growth}
\usage{
percent_change(x, dp = 1, form = "f", ...)
}
\arguments{
\item{x}{numeric representation of a percentage}

\item{dp}{the number of decimal places to display in the output}

\item{form}{Same as the \code{format} parameter in \code{\link{formatC}}}

\item{...}{Additional options for formatC}
}
\value{
A character representation of \code{x} as a percentage
}
\description{
A modification of the \code{\link[scales]{percent}} function from \code{scales}, which
gives better control over the decimal places displayed
}
\examples{
percent(0.01)
percent(c(0.001, 0.23423452, 0.2, 1.124, 1213.34))

Compare to scales implementation:
scales::percent(c(0.001, 0.23423452, 0.2, 1.124, 1213.34))
}
