% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{ppt_png}
\alias{ppt_png}
\title{Save of PNG for PowerPoint}
\usage{
ppt_png(p, filename, w = 22, h = 12, fam = "Calibri", pm = c(0.1, 0.1,
  0.25, 0.1), mvl_foot = TRUE, mvl_foot_text = "",
  mvl_foot_colour = mvl_grey, mvl_foot_size = 0.8)
}
\arguments{
\item{mvl_foot_size}{}
}
\description{
Saves a ggplot object as a PNG that is suitable for a Marketview
report/presentation in ppt.
}

