% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{source_dir}
\alias{source_dir}
\title{Source all scripts in a directory}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{The directory containing the scripts}

\item{trace}{Print status to the console?}

\item{...}{Other options}
}
\description{
Source all scripts in a directory
}
