% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread2.R
\name{spread2}
\alias{spread2}
\title{Spread a key-value pair across multiple columns.}
\usage{
spread2(data, key, value, ..., fill = NA, convert = FALSE, drop = TRUE,
  sep = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{key}{Column names or positions. This is passed to tidyselect::vars_pull().}

\item{value}{These arguments are passed by expression and support quasiquotation (you can unquote column names or column positions).}

\item{...}{Which additional columns in \code{data} should be kept in the output?}

\item{fill}{If set, missing values will be replaced with this value. Note that there are two types of missingness in the input: explicit missing values (i.e. NA), and implicit missings, rows that simply aren't present. Both types of missing value will be replaced by fill.}

\item{convert}{If TRUE, type.convert() with asis = TRUE will be run on each of the new columns. This is useful if the value column was a mix of variables that was coerced to a string. If the class of the value column was factor or date, note that will not be true of the new columns that are produced, which are coerced to character before type conversion.}

\item{drop}{If FALSE, will keep factor levels that don't appear in the data, filling in missing combinations with fill.}

\item{sep}{If NULL, the column names will be taken from the values of key variable. If non-NULL, the column names will be given by "<key_name><sep><key_value>".}
}
\description{
This is a wrapper for \code{tidyr::spread}, which has an additional argument \code{...}, where you can list any other attributes other than \code{key} and \code{value} that should be kept in the output. All other attributes are dropped.
}
