% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{ssodm.plot}
\alias{ssodm.plot}
\title{This function plots the Off diagonal mass of an observed Transition Matrix compared to a set of null TMs}
\usage{
ssodm.plot(tm.obs, tm.null, sort.by.sig = F, rescale = F, plot.title = NA,
  highlight.tfs = NA)
}
\arguments{
\item{tm.obs}{The observed transition matrix}

\item{tm.null}{A list of null transition matrices}
}
\description{
This function plots the Off diagonal mass of an observed Transition Matrix compared to a set of null TMs
}
\examples{
example1
}
\keyword{keywords}

