% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_map}
\alias{theme_map}
\title{Marketview ggplot map theme.}
\usage{
theme_map(base_size = 12, base_family = "calb", plain_family = "cal")
}
\arguments{
\item{base_size}{The base size for fonts}

\item{base_family}{The base family for fonts}

\item{plain_family}{The plain family for fonts (used in legend text)}
}
\value{
A theme object to be added to ggplot objects
}
\description{
A theme for maps that looks okay in a Marketview ppt report.
Similar to \code{\link{theme_mvl}} but it has a blue background, no axis labels,
and a floating legend.
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x = 1:5, y = 1:5)) + geom_point(aes(x, y))
p + theme_map()
}

