% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_map_minimal}
\alias{theme_map_minimal}
\title{Marketview ggplot map theme (minimal).}
\usage{
theme_map_minimal(base_size = 12, base_family = "calb")
}
\arguments{
\item{base_size}{The base size for fonts}

\item{base_family}{The base family for fonts}
}
\value{
A theme object to be added to ggplot objects
}
\description{
A minimal theme for maps that is more suitable for faceted plots.
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x = 1:5, y = 1:5)) + geom_point(aes(x, y))
p + theme_map_minimal()
}

