% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_mcd}
\alias{theme_mcd}
\title{McDonald's ggplot theme.}
\usage{
theme_mcd(base_size = 11, base_family = "cenb", plain_family = "cen",
  text_colour = "black")
}
\arguments{
\item{base_size}{The base size for fonts}

\item{base_family}{The base family for fonts}
}
\value{
A theme object to be added to ggplot objects
}
\description{
A common theme for McDonalds plots.
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x = 1:5, y = 1:5)) + geom_point(aes(x, y))
p + theme_mcd()
}

