% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_mvl}
\alias{theme_mvl}
\title{Marketview ggplot theme.}
\usage{
theme_mvl(base_size = 11, base_family = "hnb", plain_family = "hn",
  text_colour = "black")
}
\arguments{
\item{base_size}{The base size for fonts}

\item{base_family}{The base family for fonts}

\item{plain_family}{The plain family for fonts (used in legend text)}
}
\value{
A theme object to be added to ggplot objects
}
\description{
A minimal theme for plots that looks okay in a Marketview ppt report.
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x = 1:5, y = 1:5)) + geom_point(aes(x, y))
p + theme_mvl()
}

