% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tld_comparison_plot.R
\name{tld_comparison_plot}
\alias{tld_comparison_plot}
\title{Comparison Plot of TLD Item Performance}
\usage{
tld_comparison_plot(con, product_tbl, group_name, align_weeks = FALSE,
  index = FALSE)
}
\arguments{
\item{index}{}
}
\description{
Comparison Plot of TLD Item Performance
}
