% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tld_customer_data.R
\name{tld_customer_data}
\alias{tld_customer_data}
\title{Get Matched TLD Customer Data}
\usage{
tld_customer_data(con, codes, start_week, end_week, keep_table = FALSE,
  tablename = NULL)
}
\arguments{
\item{con}{ODBC connection to bespoke}

\item{codes}{Set of codes for which to get the customers}

\item{start_week}{The first promo week of the time period}

\item{end_week}{The last promo week of the time period}

\item{keep_table}{Should the customer table be kept on the database? Use this if you don't want to re-run the data}

\item{tablename}{Name of the permanent table in the database, if you want to keep it}
}
\description{
Get Matched TLD Customer Data
}
