% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tld_item_friends.R
\name{tld_item_friends}
\alias{tld_item_friends}
\title{Most Common Item Associations}
\usage{
tld_item_friends(con, codes, start_week, end_week)
}
\arguments{
\item{con}{ODBC connection to bespoke database}

\item{codes}{The menu_item_no codes of the items to consider}

\item{start_week}{The first promo week of the time period to consider}

\item{end_week}{The last promo week of the time period to consider}
}
\description{
What are the best friends of a MCD tld item? Gets the number of items, total spending,
and total transactions for all items that have been bought in the same transaction
as the given items.
}
