% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tld_sales_summary.R
\name{tld_sales_summary}
\alias{tld_sales_summary}
\title{Get the Weekly Sales for a TLD Item}
\usage{
tld_sales_summary(con, codes, start_week, end_week, group_name = NULL)
}
\arguments{
\item{con}{ODBC connection to bespoke}

\item{codes}{The menu_item_no codes of the items}

\item{start_week}{The first promo week of the time period}

\item{end_week}{The last promo week of the time period}

\item{group_name}{Optional name attribute to add to the output data}
}
\description{
Get the Weekly Sales for a TLD Item
}
