% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{trim_ws}
\alias{trim_ws}
\title{Remove whitespace from character vectors}
\usage{
trim_ws(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Text vector with leading and trailing whitespace removed
}
\description{
Remove whitespace from character vectors
}
\examples{
trim_ws(c("test_name", "text_artifact       "))
}

