% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval_functions.R
\name{vec_sql_string}
\alias{vec_sql_string}
\title{Parse a String Vector for SQL Queries}
\usage{
vec_sql_string(vec)
}
\arguments{
\item{vec}{A vector of strings to convert to a SQL list}
}
\value{
A string containing the values in \code{vec} in SQL-list format
}
\description{
Parse a String Vector for SQL Queries
}

