% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{zoom_definition}
\alias{zoom_definition}
\title{Define a Zooming Function for Plotting Maps.}
\usage{
zoom_definition(shp_dat, x = "long", y = "lat")
}
\arguments{
\item{shp_dat}{A shapefile on which to base the plot window}

\item{x}{The name of the attribute in shp_dat which describes the horizontal coordinates}

\item{y}{The name of the attribute in shp_dat which describes the vertical coordinates}
}
\description{
For a shapefile in NZTM format return a 'zoom' function (a wrapper to coord_fixed()) which displays a square area centred on the shapefile.
}
\examples{
akl_zoom <- zoom_definition(akl_cau_13.df)
ggplot()+
geom_polygon(data = nz_tla_13.df,
aes(long, lat, group = group, fill = TLA == 76))+
akl_zoom()
}

