% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{zoom_place}
\alias{zoom_place}
\title{Define a Zooming for Maps, Based on a Named Area.}
\usage{
zoom_place(place_name, margin = 20000, expand = FALSE, ratio = 1)
}
\arguments{
\item{place_name}{Name of the City/Town/Suburb/etc to centre the map on}

\item{margin}{The margin that should be added to the plot area}

\item{expand}{If TRUE, adds a small expansion factor to the limits to ensure that data and axes don't overlap. If FALSE, limits are taken exactly from the data or xlim/ylim.}

\item{ratio}{The aspect ratio (x/y) that the resultant plot area should have}
}
\description{
A wrapper function for zoom_definintion() which refers to places from mvldata::osm_places.df to define a bounding box for the map
}
\examples{
ggplot()+
geom_polygon(data = nz_tla_13.df,
aes(long, lat, group = group, fill = TLA == 76))+
zoom_place("Auckland")
}
