% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_fwhen.R
\name{case_fwhen}
\alias{case_fwhen}
\title{Vectorised if returning an ordered result}
\usage{
case_fwhen(...)
}
\arguments{
\item{...}{A sequence of two-sided formulas. The left hand side (LHS)
determines which values match this case. The right hand side (RHS)
provides the replacement value.

The LHS must evaluate to a logical vector. Each logical vector can
either have length 1 or a common length. All RHSs must evaluate to
the same type of vector.

These dots are evaluated with \link[rlang:dots_list]{explicit splicing}.}
}
\description{
This function allows you to vectorise multiple if and else if statements. It is an R equivalent of the SQL CASE WHEN statement.
An update from the dplyr implementation as it returns the result as an ordered factor in the order given by the formulae
}
