% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{data_import}
\alias{data_import}
\title{Import the processed data for a project}
\usage{
data_import(data_dir = file.path("data", "processed"),
  r_dir = file.path("R"))
}
\arguments{
\item{data_dir}{The directory where the processed data lives}

\item{r_dir}{The directory where the processing scripts live}
}
\value{
imports the processed data into the global environment
}
\description{
Import the processed data for a project
}
