% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{data_import}
\alias{data_import}
\title{Import the processed data for a project}
\usage{
data_import(base_dir, data_name = paste(gsub(".*/(?!$)|/$", "", base_dir, perl
  = TRUE), "data.RData"))
}
\arguments{
\item{data_name}{}
}
\description{
Import the processed data for a project
}

