% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mvl.R
\name{pal_mvl}
\alias{pal_mvl}
\title{Marketview Palette}
\usage{
pal_mvl(palette = "Leaf", direction = 1)
}
\arguments{
\item{palette}{Name of the palette to use}

\item{direction}{Set the order of the colours in the scale}
}
\description{
The \code{Marketview} palettes provide sequential, diverging and qualitative
colour schemes in line with the Marketview branding guides.
}
\section{Palettes}{

The following palettes are available for use with these scales:
\describe{
  \item{Diverging}{LeafPlum, TealRed, RedStone, SkyStone, PlumStone, LeafStone, RoyalStone}
  \item{Qualitative}{mvl, mvl_classic, mcd, mcd2}
  \item{Sequential}{Leaf, Plum, Teal, Blue, Sky, Citrus, Stone,
     Navy, Orange, Red, Green, Purple}
}
If there are others that you want to be implemented, let me (Bert) know
}

