% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl_functions.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot object.}
\usage{
save_plot(p, write_dir, name, fig_num = 1, height = 700 * 4, width = 850 *
  4, res = 360, mvl_footer = FALSE)
}
\arguments{
\item{p}{A ggplot object}

\item{write_dir}{The directory in which the figure should be saved}

\item{name}{A character name that the saved object will have}

\item{fig_num}{A figure number that preceeds the name}

\item{height}{The height of the png}

\item{width}{The width of the png}

\item{res}{The resolution of the png}

\item{mvl_footer}{Should a Marketview footer image be included?}
}
\value{
Saves the plot to file
}
\description{
A function to save a plot using dimensions and resolution that are
suitable for a Marketview report.
}

