% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tld_comparison_plot.R
\name{tld_comparison_plot}
\alias{tld_comparison_plot}
\title{Comparison Plot of TLD Item Performance}
\usage{
tld_comparison_plot(con, product_tbl, group_name = "", align_weeks = FALSE,
  index = FALSE, promo_week = TRUE)
}
\arguments{
\item{con}{ODBC connection to bespoke}

\item{product_tbl}{Table of products to consider. Attributes should be: codes | item_name | start_week | end_week}

\item{group_name}{Name for the plot title}

\item{align_weeks}{Should the weeks be standardised to 'Week 1', 'Week 2', etc? Otherwise uses seqpromo_week}

\item{index}{Should AWUs be indexed to the first week?}
}
\description{
Comparison Plot of TLD Item Performance
}
